Move Texture2DExtensions into Extensions namespace
