+++ Assets/Scripts/Clipboard/Clipboard.cs
@@ -1,8 +1,10 @@
 using PAC.Animation;
 using PAC.DataStructures;
 using PAC.Drawing;
+using PAC.Extensions;
 using PAC.Files;
 using PAC.Layers;
+
 using UnityEngine;
 
 namespace PAC.Clipboard
