+++ Assets/Scripts/Files/FileManager.cs
@@ -1,10 +1,14 @@
 using System.Collections.Generic;
 using System.IO;
+
 using PAC.Animation;
+using PAC.Extensions;
 using PAC.Input;
 using PAC.Layers;
 using PAC.UI;
+
 using SFB;
+
 using UnityEngine;
 using UnityEngine.Events;
