+++ Assets/Scripts/Layers/Layer.cs
@@ -1,10 +1,13 @@
 using System;
 using System.Collections.Generic;
 using System.Runtime.Serialization;
+
 using PAC.Animation;
 using PAC.Colour;
 using PAC.DataStructures;
+using PAC.Extensions;
 using PAC.Json;
+
 using UnityEngine;
 using UnityEngine.Events;
