+++ Assets/Scripts/Layers/LayerManager.cs
@@ -1,11 +1,13 @@
 using System.Collections.Generic;
 using System.Linq;
+
 using PAC.Animation;
 using PAC.Colour;
 using PAC.Extensions;
 using PAC.Files;
 using PAC.UI;
 using PAC.UndoRedo;
+
 using UnityEngine;
 using UnityEngine.Events;
