+++ Assets/Scripts/Layers/NormalLayer.cs
@@ -1,9 +1,12 @@
 using System.Collections.Generic;
 using System.Runtime.Serialization;
+
 using PAC.Animation;
 using PAC.Colour;
 using PAC.DataStructures;
+using PAC.Extensions;
 using PAC.Json;
+
 using UnityEngine;
 
 namespace PAC.Layers
